Use timedelta for intervals
The reasoning before was that Postgres seems to have months and years
in its intervals which can't be correctly converted to days (because
of leaps years and differing day count in months).

Switching to binary showed that Postgres itself might not store all
this information exactly either and use some approximations (like a
month is 30 days). Comparing with asyncpg also seems to confirm
this. The new code also uses iso_8601 for IntervalStyle setting, which
is much easier to parse.