Attempt a graceful shutdown when closing
When a closing the connection normally, we now send a Terminate
message to the server before closing the stream.

Notice that this will not work if the connection is forcibly closed
due to an exception. We could likely gracefully terminate the
connection in that case as well, but it would be more complicated, and
I did not bother for now since it's not that important anyways.