Synchronize transactions across tasks
Concurrent transactions in different tasks could cause issues
before. We now keep a context variable which tells us if we're inside
a transaction block, in the same task or not.

If we're not, we start the transaction after acquiring a lock. If
we're inside another transaction in the same task, we create a
savepoint as before.