Allow tests to be run as root
In certain environments (such as github actions), we might need to run
the tests as root, but pg_ctl refuses to be run as the root user. We
now detect this, and use sudo to run pg_ctl as the postgres user, and
also chown the temp data directory to be owned by postgres.