Delete old test db before creating new one
This approach works better than before where we deleted the test
databases after the yield statement in the cleanup. The old method ran
into lots of issues if a connection failed, because the fixture task
was being canceled. In many cases, the database was not dropped,
causing the next case to fail when it tried to create the test
database.