Re-run failed tests in github action
When running the github action, sometimes the database connection
breaks. We want to retry these tests. We're using the
pytest-rerunfailures plugin for this. Ideally, we should pass
`--only-rerun OperationalError` so that only OperationalError
exceptions cause retries, but the plugin seems to have a bug and using
that option, causes internal errors.