Discard broken connections in the pool
Broken connections raise OperationalError exceptions, leaving no place
to handle them properly. The Pool objects now register a callback with
the Connections they manage, which will be called, instead of an
exception, when a broken connection is detected. These will then be
discarded in the pool and replaced by another connection later.